requirements.txt
appdirs==1.4.4
argcomplete==3.2.1
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
importlib-resources==6.1.1
Pillow==10.1.0
pypng==0.20220715.0
pyserial==3.5
python-barcode==0.15.1
python-decouple==3.8
python-escpos==3.1
pytz==2023.3.post1
pyusb==1.2.1
PyYAML==6.0.1
qrcode==7.4.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
